aibolit/patterns/mutable_class/mutable_class.py
# The MIT License (MIT)
#
# Copyright (c) 2020 Aibolit
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included
# in all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

from typing import List
from aibolit.types_decl import LineNumber


class MutableClass:

    def __init__(self):
        pass

    def value(self, filename) -> List[LineNumber]:
        """
        Finds if a class is mutable, i.e., have non-final fields
        :param filename: filename to be analyzed
        :return: list of LineNumber with the class declaration lines
        @todo #139:30min Implement mutable_class pattern
         If we see a class that has at least one non-final attribute,
         it's a mutable class. The method must return a list of line numbers of
         the classes in file which have this condition. After implementation,
         activate tests in test_mutable_class.py
        """
        return []
